scripts/requirements.txt
tqdm
openai
google-generativeai
python-dotenv

